requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
django-otp>=1.3
django-two-factor-auth>=1.17
Pillow>=10.0
